pyproject.toml
[project]

name = "apple_pass_creator"
version = "0.1.0"

dependencies = [
    "passbook",
    "fastapi",
    "uvicorn",
    "python-multipart",
    "edutap.wallet_apple"
]
